requirements.txt
SQLAlchemy==1.4.47
python-telegram-bot==21.0.1
python-dotenv==1.0.1